requirements.txt
pynubank
fastapi
pydantic-settings
uvicorn[standard]
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib